Need to use long longs for calculations, or we hit overflow easily.
E.g., when calculating 800ms delay on 8MHz.

+++ include/delay_time.hpp
@@ -23,11 +23,11 @@ class TimeDelay
 public:
     inline static void delay_us(unsigned us)
     {
-        cycle_delayer::delay(freq * us / 1000000UL);
+        cycle_delayer::delay((unsigned long long)freq * us / 1000000UL);
     }
 
     inline static void delay_ms(unsigned ms)
     {
-        cycle_delayer::delay(freq * ms / 1000UL);
+        cycle_delayer::delay((unsigned long long)freq * ms / 1000UL);
     }
 };
